full_wrap/requirements.txt
stable-baselines3>=2.1.0
numpy==1.24.4
torch>=2.0.0
gymnasium>=0.28.1
matplotlib>=3.5.0
pandas>=1.3.0
scipy>=1.7.0
matlabengine==24.2.0
optuna>=3.3.0
tqdm>=4.65.0 